Backend/requirements.txt
fastapi
uvicorn
pandas
supabase
python-dotenv
openpyxl
bleach
pytz
requests
pydantic